[EDELIVERY-7779] customize the total number on the client; delete commented code

+++ Domibus-MSH/src/test/java/eu/domibus/core/message/UserMessageLogDaoTest.java
@@ -666,7 +666,7 @@ public class UserMessageLogDaoTest {
         }};
 
         // WHEN
-        int result = userMessageLogDao.countEntries(filters);
+        long result = userMessageLogDao.countEntries(filters);
 
         // THEN
         Assert.assertEquals("Should have returned the correct count when filters provided", 4, result);
@@ -680,7 +680,7 @@ public class UserMessageLogDaoTest {
         }};
 
         // WHEN
-        int result = userMessageLogDao.countEntries(new HashMap<>());
+        long result = userMessageLogDao.countEntries(new HashMap<>());
 
         // THEN
         Assert.assertEquals("Should have returned the total count when no filters provided", 7, result);
